Templated the global heap allocator
